% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DSC.R
\name{calculate_DSC}
\alias{calculate_DSC}
\title{Dice Similarity Coefficient Calculation}
\usage{
calculate_DSC(template0, template1, sides = c("right", "left"))
}
\arguments{
\item{template0}{Initial template mask. Right lung = 1, left lung = 2, non-lung = 0}

\item{template1}{New template mask. Right lung = 1, left lung = 2, non-lung = 0}

\item{sides}{Calculate DSC on right and/or left lungs. Overall DSC is given by default.}
}
\value{
DSC on whole lung, right lung and left lung.
}
\description{
Calculate the Dice similarity coefficient (DSC) for two iterations of lung templates
}
