% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse_body.R
\name{coarse_body}
\alias{coarse_body}
\title{Coarse Masking of Image from Z-slicding}
\usage{
coarse_body(x, keep_slices = 20)
}
\arguments{
\item{x}{Binary \code{antsImage}}

\item{keep_slices}{number of slices to keep if needed}
}
\value{
Array or \code{antsImage} object
}
\description{
Calculates a Hull at each slice then stacks them
}
