% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_body.R
\name{ellipsoid_body}
\alias{ellipsoid_body}
\title{Coarse Masking of Image from Z-slicding}
\usage{
ellipsoid_body(x)
}
\arguments{
\item{x}{Binary \code{antsImage}}
}
\value{
Array or \code{antsImage} object
}
\description{
Calculates an Ellipsoid at each slice then stacks them
}
