% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Create a lung template from warped images and composite transformations}
\usage{
get_template(folder_warp, folder_comp, outfile, gradientStep = 0.2,
  mask = TRUE, verbose = TRUE)
}
\arguments{
\item{folder_warp}{Folder path for warped images}

\item{folder_comp}{Folder path for composite transformations}

\item{outfile}{File path to save new template}

\item{gradientStep}{Gradient step size}

\item{mask}{Logical statement. TRUE if outfile should be binary.}

\item{verbose}{Print output messages}
}
\value{
Template and its filename
}
\description{
Create a lung template from warped images and composite transformations
}
