% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian_lung.R
\name{jacobian_lung}
\alias{jacobian_lung}
\title{Calculate Jacobian Determinant Image for the Lungs}
\usage{
jacobian_lung(right_transformation, left_transformation, mask,
  doLog = TRUE, geom = FALSE, relative = TRUE)
}
\arguments{
\item{right_transformation}{Transformations for the right lung}

\item{left_transformation}{Transformations for the left lung}

\item{mask}{Fixed mask (or standard lung template mask)}

\item{doLog}{Return the log Jacobian}

\item{geom}{Use the geometric Jacobian calculation}

\item{relative}{Returen the relative Jacobian (Jacobian divided by mean Jacobian)}
}
\value{
Jacobian Image
}
\description{
This function calculates the Jacobian determinant image for the left and right lungs separately.
}
