% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_reg.R
\name{lung_reg}
\alias{lung_reg}
\title{CT Lung Registration}
\usage{
lung_reg(base, follow, base_mask, follow_mask, add_1025 = TRUE,
  interpolator = "LanczosWindowedSinc", typeofTransform = "SyN", ...)
}
\arguments{
\item{base}{Baseline scan - filename or \code{antsImage}}

\item{follow}{Follow up scan - filename or \code{antsImage}
Registered to the baseline scan.}

\item{base_mask}{Baseline mask - binary image for baseline.}

\item{follow_mask}{Follow up mask - binary image for follow up scan}

\item{add_1025}{Should 1025 be added to make values non-negative (recommended
for CT)}

\item{interpolator}{interpolation in baseline space,
passed to \code{\link{registration}}}

\item{typeofTransform}{transformation to use, passed to \code{\link{registration}}}

\item{...}{additional options to pass to \code{\link{registration}}}
}
\value{
List of output, same as \code{\link{registration}}, with transforms
and output files, as well as an indicator for \code{add_1025}, the
value added to the data, as well as the baseline scan
}
\description{
Wrapper of regsitration function for CT scans of the lung
}

