% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_lung.R
\name{partition_lung}
\alias{partition_lung}
\title{Partition the lung}
\usage{
partition_lung(img, kernel_size = c(30, 30, 30), kernel_stride = c(30,
  30, 30), centroid = TRUE)
}
\arguments{
\item{img}{Lung scan to be partitioned. Can be in 3D matrix or ANTs image file format.}

\item{kernel_size}{Size of the kernel, in voxel units of width, depth, and height. Must be c(3,3,3) or greater. Default: c(30,30,30)}

\item{kernel_stride}{Stride (or spacing) between kernels, in voxel units, for width, depth, and height. If kernel_stride = kernel_size, the partitions are non-overlapping. If stride = c(1,1,1), then each voxel is returned.}

\item{centroid}{Logical. If true, output includes the centroids of each partition.}
}
\value{
Matrix of x, y, and z coordinates for each partition
}
\description{
Partition the lung
}
