% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiomics_partition.R
\name{radiomics_partition}
\alias{radiomics_partition}
\title{Calculate radiomic features on the partitioned 3D lung}
\usage{
radiomics_partition(img, mask, mask_values = c(1, 2),
  featuresFirst = c("mean", "sd", "skew", "kurtosis", "min", "q1",
  "median", "q3", "max", "energy", "rms", "uniformity", "entropy"),
  featuresSpatial = c("mi", "gc", "fd"), partition = NULL,
  kernel_size = c(30, 30, 30), kernel_stride = c(30, 30, 30),
  threshold = 1000, tidy = TRUE)
}
\arguments{
\item{img}{CT scan in ANTs image file format}

\item{mask}{Mask of CT scan in ANTs image file format}

\item{mask_values}{Values of mask to use for radiomic feature calculation}

\item{featuresFirst}{First level radiomic features to calculate}

\item{featuresSpatial}{Spatial radiomic features to calculate}

\item{partition}{Matrix of x, y, and z coordinates for each partition from partition_lung. If null, partition_lung is called.}

\item{kernel_size}{(If partition is null) Size of the kernel, in voxel units of width, depth, and height. Must be c(3,3,3) or greater. Default: c(30,30,30)}

\item{kernel_stride}{(If partition is null) Stride (or spacing) between kernels, in voxel units, for width, depth, and height. If kernel_stride = kernel_size, the partitions are non-overlapping. If stride = c(1,1,1), then each voxel is returned.}

\item{threshold}{Number of non-missing voxels needed to calculate radiomic features in each partition.}

\item{tidy}{Logical. If true, outputs a tidy dataframe with results. If false, outputs nested loop.}
}
\value{
Values from selected features for both left and right lungs
}
\description{
Calculate radiomic features on the partitioned 3D lung
}
