% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiomics_slice.R
\name{radiomics_slice}
\alias{radiomics_slice}
\title{Calculate radiomic features on the each 2D slice of the whole 3D lung, left and right lungs separately}
\usage{
radiomics_slice(img, mask, mask_values = c(1, 2), plane = "axial",
  featuresFirst = c("mean", "sd", "skew", "kurtosis", "min", "q1", "median",
  "q3", "max", "energy", "rms", "uniformity", "entropy"),
  featuresSpatial = c("mi", "gc", "fd"), tidy = TRUE)
}
\arguments{
\item{img}{CT scan in ANTs image file format}

\item{mask}{Mask of CT scan in ANTs image file format}

\item{mask_values}{Values of mask to use for radiomic feature calculation}

\item{plane}{One of: axial, coronal, sagittal}

\item{featuresFirst}{First level radiomic features to calculate}

\item{featuresSpatial}{Spatial radiomic features to calculate}

\item{tidy}{Logical. If true, outputs a tidy dataframe with results. If false, outputs nested loop.}
}
\value{
Radiomic values from every slice in both the left and right lungs
}
\description{
Calculate radiomic features on the each 2D slice of the whole 3D lung, left and right lungs separately
}
