% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiomics_spatial.R
\name{radiomics_spatial}
\alias{radiomics_spatial}
\title{Calculate spatial radiomic features on a 2D or 3D array}
\usage{
radiomics_spatial(data, features = c("mi", "gc"))
}
\arguments{
\item{data}{Any 2D or 3D image (as matrix or array) to calculate spatial features}

\item{features}{= spatial radiomic features to calculate}
}
\value{
Values from selected features
}
\description{
Calculate spatial radiomic features on a 2D or 3D array
}
