% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_scan.R
\name{reduce_scan}
\alias{reduce_scan}
\title{Reduction of Scan Dimensions from Mask}
\usage{
reduce_scan(img, mask)
}
\arguments{
\item{img}{Image to reduce}

\item{mask}{Mask to reduce image by}
}
\value{
List of antsImage objects for reduced image and mask
}
\description{
Reduces the dimensions of a scan based on extent of
a mask. Combines
\code{\link{getEmptyImageDimensions}} and
\code{\link{applyEmptyImageDimensions}}.
}
