% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_airway.R
\name{segment_airway}
\alias{segment_airway}
\title{Segmentation of airways from a CT scan}
\usage{
segment_airway(img, lung_air_mask, verbose = TRUE)
}
\arguments{
\item{img}{Original CT scan}

\item{lung_air_mask}{Mask of original CT scan from segment_lung_airway}

\item{verbose}{Print diagnostic messages}
}
\value{
Mask of airways only
}
\description{
Segmentation of airways from a CT scan
}
