% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_human.R
\name{segment_human}
\alias{segment_human}
\title{Segment Human from CT scans}
\usage{
segment_human(img, adder = 1025, lthresh = -300, verbose = TRUE)
}
\arguments{
\item{img}{Filename or \code{antsImage}}

\item{adder}{amount to be added to the image to make non-zero}

\item{lthresh}{lower threshold for the image}

\item{verbose}{Print diagnostic messages}
}
\value{
List of smoothed image, masked smoothed image, body
}
\description{
Segment Human from a non-contrast CT scan
}
