% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_lung_lr.R
\name{segment_lung2}
\alias{segment_lung2}
\title{Lung Segmentation}
\usage{
segment_lung2(img, lthresh = -300, verbose = TRUE)
}
\arguments{
\item{img}{CT scan in ANTs or NIfTI file format}

\item{lthresh}{Threshold used to find the lung and airways,
from the CT. Default: -300 HU.}

\item{verbose}{Print diagnostic messages.}
}
\value{
Lung mask in ANTs file format. Right lung = 1, left lung = 2, non-lung = 0
}
\description{
This function segments the right and left lungs from a CT scan.
}
