% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_lung_airway.R
\name{segment_lung_airway}
\alias{segment_lung_airway}
\title{Segmentation of lung and airway from a CT scan}
\usage{
segment_lung_airway(img, lthresh = -300, verbose = TRUE)
}
\arguments{
\item{img}{Original CT scan}

\item{lthresh}{Threshold used to find the lung and airways,
from the original CT scan. Anything below this threshold will
be used for the initial scan. Default: -300 HU.}

\item{verbose}{Print diagnostic messages.}
}
\value{
Lung mask of lung and airways from original CT scan.
Mask has values = 1, background = 0.
}
\description{
Segmentation of lung and airway from a CT scan
}
