% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_lung_lr.R
\name{segment_lung_lr}
\alias{segment_lung_lr}
\title{Segmentation of left and right lungs from a CT scan}
\usage{
segment_lung_lr(img, lthresh = -300, verbose = TRUE)
}
\arguments{
\item{img}{Original CT scan}

\item{lthresh}{Threshold used to find the lung and airways,
from the original CT scan. Anything below this threshold will
be used for the initial scan. Default: -300 HU.}

\item{verbose}{Print diagnostic messages.}
}
\value{
Lung mask, with left and right designation, of original CT scan.
Right lung has values = 1, left lung has values = 2, left/right lung overlap
(i.e. couldn't distinguish the left and right lungs) has values = 3, non-lung = 0.
}
\description{
Segmentation of left and right lungs from a CT scan
}
