% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Lung Template Creation}
\usage{
get_template(
  folder_warp,
  folder_comp,
  sides = c("right", "left"),
  gradientStep = 0.2,
  mask = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{folder_warp}{Folder path for warped images}

\item{folder_comp}{Folder path for composite transformations}

\item{sides}{Do both left and right or only one?}

\item{gradientStep}{Gradient step size}

\item{mask}{Logical statement. TRUE if template should be binary.}

\item{verbose}{Print output messages}
}
\value{
New Template. Right lung = 1, left lung = 2, non-lung = 0.
}
\description{
Create a new lung template from warped images and composite transformations. If the DSC < 0.99, more iterations should be performed.
}
