% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Create a lung template from warped images and composite transformations}
\usage{
get_template(folder_warp, folder_comp, sides = c("left", "right"),
  outprefix = NULL, gradientStep = 0.2, mask = TRUE,
  verbose = TRUE)
}
\arguments{
\item{folder_warp}{Folder path for warped images}

\item{folder_comp}{Folder path for composite transformations}

\item{sides}{Do both left and right or only one?}

\item{outprefix}{Folder and file path to save new template (don't include extension)}

\item{gradientStep}{Gradient step size}

\item{mask}{Logical statement. TRUE if outfile should be binary.}

\item{verbose}{Print output messages}
}
\value{
Template and its filename
}
\description{
Create a lung template from warped images and composite transformations
}
