% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_lung.R
\name{register_lung}
\alias{register_lung}
\title{Register lungs, with options to save warped image and composed transformation}
\usage{
register_lung(infile_template, infile_moving, template_masked = TRUE,
  mask_value = 1, typeofTransform = "SyN", outfile_warp = NULL,
  outfile_comp = NULL, out = FALSE, verbose = TRUE,
  verbose_reg = FALSE)
}
\arguments{
\item{infile_template}{File path for the template (fixed) image}

\item{infile_moving}{File path for the moving image}

\item{template_masked}{Logical. Is the template already masked?}

\item{mask_value}{Value to mask moving image, and template if necessary}

\item{typeofTransform}{Type of transformation for registration}

\item{outfile_warp}{File name to save warped image. If NULL, warped image is not saved to file.}

\item{outfile_comp}{File name to save composite transformation. If NULL, transformation is not saved to file.}

\item{out}{Logical. Should the registration object be returned?}

\item{verbose}{Print messages}

\item{verbose_reg}{Print messages, specifically related to registration function}
}
\value{
antsRegistration object, if out = TRUE
}
\description{
Register lungs, with options to save warped image and composed transformation
}
