% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_lung_mask.R
\name{register_lung_mask}
\alias{register_lung_mask}
\title{Lung Registration}
\usage{
register_lung_mask(moving_mask, fixed_mask, moving = NULL,
  sides = c("right", "left"), verbose = FALSE, typeofTransform = "SyN",
  composeTransforms = NULL, mask_interpolator = "nearestNeighbor",
  interpolator = "linear", ...)
}
\arguments{
\item{moving_mask}{Mask of moving image. Right lung = 1, left lung = 2, non-lung = 0}

\item{fixed_mask}{Mask of fixed image. Right lung = 1, left lung = 2, non-lung = 0}

\item{moving}{Moving image to apply transformation}

\item{sides}{Choose to register right and/or left lungs.}

\item{verbose}{Print diagnostic messages}

\item{typeofTransform}{Type of transform, passed to \code{\link{antsRegistration}}}

\item{composeTransforms}{Prefix of output filename to save the composed forward transformations. The prefix will add comptx.nii.gz to the end.}

\item{mask_interpolator}{Interpolator used to apply transformation to moving mask,
passed to \code{\link{antsApplyTransforms}}}

\item{interpolator}{Interpolator used to apply transformation to image,
passed to \code{\link{antsApplyTransforms}}}

\item{...}{addition arguments to pass to \code{\link{antsRegistration}}}
}
\value{
A list of warped masks, images, and transformations for
right and left lungs separately
}
\description{
This function registers the right and left lung masks to a template mask. To register to the standard lung template mask, type \code{system.file("extdata", "lung_template_mask.nii.gz", package = "lungct").}
}
