% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungreg_function.R
\name{register_lung_mask}
\alias{register_lung_mask}
\title{Lung Registration for Left and Right Lungs}
\usage{
register_lung_mask(moving_mask, fixed_mask, moving = NULL, sides = c("left",
  "right"), outprefix = NULL, verbose = FALSE,
  mask_interpolator = "nearestNeighbor", interpolator = "linear",
  save_memory = TRUE, ...)
}
\arguments{
\item{moving_mask}{Mask of moving image, left has values
1 and 3 (3 if both left and right), right has values 2 and 3}

\item{fixed_mask}{Mask of fixed image, left has values
1 and 3 (3 if both left and right), right has values 2 and 3}

\item{moving}{image of moving image, usually CT, will be masked
by \code{moving_mask}}

\item{sides}{Do both left and right or only one?}

\item{outprefix}{Path to put the transformations,
passed to \code{\link{antsRegistration}}}

\item{verbose}{Print diagnostic messages}

\item{mask_interpolator}{interpolator used to apply transformation to moving mask,
passed to \code{\link{antsApplyTransforms}}}

\item{interpolator}{interpolator used to apply transformation to image,
passed to \code{\link{antsApplyTransforms}}}

\item{save_memory}{do garbage collection and not save intermediate
files in registration.}

\item{...}{addition arguments to pass to \code{\link{antsRegistration}}}
}
\value{
A list of registrations, transformed images, each for
left and right separately
}
\description{
Lung Registration for Left and Right Lungs
}
