% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_lung.R
\name{segment_lung}
\alias{segment_lung}
\title{Segment Lungs from CT scans}
\usage{
segment_lung(img, mask = TRUE, lthresh = -300, verbose = TRUE)
}
\arguments{
\item{img}{Filename or \code{antsImage}}

\item{mask}{Should the image be masked}

\item{lthresh}{lower threshold for the image}

\item{verbose}{Print diagnostic messages}
}
\value{
List of image, lung, lung mask
}
\description{
Segment Lungs from a non-contrast CT scan
}
